add params

+++ data/view/cnf/def.go
@@ -46,13 +46,13 @@ var TypeMysqlDicMp = map[string]string{
 
 // TypeMysqlMatchMp Fuzzy Matching Types.模糊匹配类型
 var TypeMysqlMatchMp = map[string]string{
-	`^(tinyint)[(]\d+[)] unsigned`:    "uint8",
-	`^(smallint)[(]\d+[)] unsigned`:   "uint16",
-	`^(int)[(]\d+[)] unsigned`:        "uint32",
-	`^(bigint)[(]\d+[)] unsigned`:     "uint64",
-	`^(tinyint)[(]\d+[)]`:             "int8",
-	`^(smallint)[(]\d+[)]`:            "int16",
-	`^(int)[(]\d+[)]`:                 "int",
+	`^(tinyint)[(]\d+[)] unsigned`:    "int64",
+	`^(smallint)[(]\d+[)] unsigned`:   "int64",
+	`^(int)[(]\d+[)] unsigned`:        "int64",
+	`^(bigint)[(]\d+[)] unsigned`:     "int64",
+	`^(tinyint)[(]\d+[)]`:             "int64",
+	`^(smallint)[(]\d+[)]`:            "int64",
+	`^(int)[(]\d+[)]`:                 "int64",
 	`^(bigint)[(]\d+[)]`:              "int64",
 	`^(char)[(]\d+[)]`:                "string",
 	`^(enum)[(](.)+[)]`:               "string",
